@@ -1,7 +1,7 @@
 {
     "name": "craigpaul/moneris-api",
-    "description": "An easier way to consume the (truly awful) Moneris eSELECTplus API.",
-    "keywords": ["moneris", "eselectplus", "api"],
+    "description": "An easier way to consume the Moneris eCommerce Unified API.",
+    "keywords": ["moneris", "api"],
     "license": "MIT",
     "authors": [
         {
@@ -10,14 +10,14 @@
         }
     ],
     "require": {
-        "php": ">=5.6",
+        "php": "^8.0",
         "guzzlehttp/guzzle": "^6.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "~5.0",
-        "mockery/mockery": "~0.9",
+        "phpunit/phpunit": "^9.3",
+        "mockery/mockery": "^1.4",
         "symfony/var-dumper": "^3.1",
-        "fzaninotto/faker": "^1.6",
+        "fakerphp/faker": "^1.6",
         "squizlabs/php_codesniffer": "^2.7"
     },
     "autoload": {
