@@ -16,7 +16,7 @@ class CreditCardTest extends TestCase
      *
      * @return void
      */
-    public function setUp()
+    public function setUp (): void
     {
         parent::setUp();
 
