@@ -41,7 +41,7 @@ class GatewayTest extends TestCase
      *
      * @return void
      */
-    public function setUp()
+    public function setUp (): void
     {
         parent::setUp();
 
