@@ -17,7 +17,7 @@ class MonerisTest extends TestCase
      *
      * @return void
      */
-    public function setUp()
+    public function setUp (): void
     {
         parent::setUp();
 
