@@ -42,7 +42,7 @@ class ProcessorTest extends TestCase
      *
      * @return void
      */
-    public function setUp()
+    public function setUp (): void
     {
         parent::setUp();
 
